+++ packages/cli/CHANGELOG.md
@@ -1,1 +1,19 @@
-# @armit/helpers
+# @armit/cli
+
+## 0.0.2
+
+### Patch Changes
+
+- release 0.0.1
+
+- Updated dependencies []:
+  - @armit/common@0.0.2
+
+## 0.0.2-next.0
+
+### Patch Changes
+
+- release 0.0.1
+
+- Updated dependencies []:
+  - @armit/common@0.0.2-next.0