+++ packages/eslint-config-bases/CHANGELOG.md
@@ -1,1 +1,13 @@
 # @armit/eslint-config-bases
+
+## 0.0.2
+
+### Patch Changes
+
+- release 0.0.1
+
+## 0.0.2-next.0
+
+### Patch Changes
+
+- release 0.0.1